# Update optimize_model=True in llama2 chatglm2 arc examples