--- a/python/llm/example/gpu/chatglm2/generate.py
+++ b/python/llm/example/gpu/chatglm2/generate.py
@@ -44,7 +44,7 @@ if __name__ == '__main__':
     # which convert the relevant layers in the model into INT4 format
     model = AutoModel.from_pretrained(model_path,
                                       load_in_4bit=True,
-                                      optimize_model=False,
+                                      optimize_model=True,
                                       trust_remote_code=True)
     model = model.to('xpu')