… python/llm/example/gpu/gpt-j/readme.md