## Description

1. Update optimize_model=True in llama2 chatglm2 arc examples.
2. gpt-j example add `model = ipex.optimize(model.eval(), dtype="float16", inplace=True)`.

### 1. Why the change?

PR[#8855](https://github.com/intel-analytics/BigDL/pull/8855) fixed the attention optimization issue on arc.

```bash
bigdl-core-xe             2.4.0b20230831           pypi_0    pypi
bigdl-llm                 2.4.0b20230831           pypi_0    pypi
```

#### chatglm2
1st token: 157.5ms -> 157.8ms
2nd+: 23.4 ms/token -> 22.6 ms/token

#### llama2-7b
1st token:  210.8 ms -> 211.9 ms
2nd+: 26.6 ms/token -> 25.3 ms/token

#### llama2-13b
1st token: 350.1 ms ->349.9 ms
2nd+:  45.1 ms/token -> 43.5 ms/token

#### gpt-j `ipex.optimize`
1st token:  200.8 ms -> 178.4ms
2nd+:  40.9 ms/token -> 22.3 ms/token

### 4. How to test?
- [ ] Unit test
- [x] Manual test